Add location to state result as described in README.
At the end of the process of bike provisioning, the bike location should be included into the FSM state: now included.

Next: JSON should be returned in REST API calls, not just a string with the state.

To improve: it is the external service that should provide the resulting bike location in `OpCompleted`: currently it is internally autogenerated randomly.
Should create a stub external service to make things nore clear on who does what.