Fixed (arguably) dependency between pagage bikes and Procurement.
DSM is fully triabgular now, but I am not fully convinced this last one is the right change because now `Procurement` messages are in package `bikes`. `AdaptedReply` can now be removed because message types are now in a third class, package bikes.
Before, I had Bike depend on Procurement for the messages tha tit needed to send to Procurement and viceversa. That's why AdaptedReply was introduced. Now that Procurement message classes are in a third place, adaptation is no longer needed.