Add two websocket actions: kill member, kick fsm.
Now by clicking with the mouse on a member node in the monitor.html page, that node will leave the cluster.
If an entity in ErrorState is clicked, then it will be kicked to continue, it will get unstuck.

On Postman, the inventory GET call now saves in the environment either the last entity in InitState or in ErrorState, so that they can be kicked by POST Restart or PUT Kick easily.
Also in Postman, the action Get Tree D3.js should now show the cluster as in monitor.html.